requirements.txt
langchain>=0.0.345
openai>=1.3.7
pandas>=2.1.3
python-dotenv>=1.0.0
python-frontmatter>=1.0.1
redis>=5.0.1
